\name{ghana}
\alias{ghana}
\docType{data}
\title{
Ranking data
}
\description{
This is a real dataset containing information on 5 object variables describing the properties of 13 different sweet potato varieties. In addition, the dataset contains partial
rankings made by men and women from Ghana.
}
\usage{data("ghana")}
\format{
  A list with three dataframes. The first consists of the rankings made by men, the second consists of the rankings made by women and the third contain the object variables.
}
\details{
Contains a subset of the data used in the Hermes et al. (2024) paper.
}
\source{
Data from the Hermes et al. (2024) paper is based on Moyo et al. (2021).
}
\references{
1. Hermes, S., van Heerwaarden, J., and Behrouzi, P. (2024). Joint Learning from Heterogeneous Rank Data. arXiv preprint, arXiv:2407.10846 \cr
2. Moyo, M., R. Ssali, S. Namanda, M. Nakitto, E. K. Dery, D. Akansake, J. Adjebeng-Danquah, J. van Etten,
K. de Sousa, H. Lindqvist-Kreuze, et al. (2021). Consumer preference testing of boiled sweetpotato using
crowdsourced citizen science in Ghana and Uganda. Frontiers in Sustainable Food Systems 5, 620363. \cr
}
\examples{
data(ghana)
}
\keyword{datasets}
